+++ resource-manager/loadtestservice/2022-12-01/client.go
@@ -6,7 +6,6 @@ package v2022_12_01
 import (
 	"fmt"
 
-	"github.com/hashicorp/go-azure-sdk/resource-manager/loadtestservice/2022-12-01/loadtest"
 	"github.com/hashicorp/go-azure-sdk/resource-manager/loadtestservice/2022-12-01/loadtests"
 	"github.com/hashicorp/go-azure-sdk/resource-manager/loadtestservice/2022-12-01/quotas"
 	"github.com/hashicorp/go-azure-sdk/sdk/client/resourcemanager"
@@ -14,18 +13,11 @@ import (
 )
 
 type Client struct {
-	LoadTest  *loadtest.LoadTestClient
 	LoadTests *loadtests.LoadTestsClient
 	Quotas    *quotas.QuotasClient
 }
 
 func NewClientWithBaseURI(sdkApi sdkEnv.Api, configureFunc func(c *resourcemanager.Client)) (*Client, error) {
-	loadTestClient, err := loadtest.NewLoadTestClientWithBaseURI(sdkApi)
-	if err != nil {
-		return nil, fmt.Errorf("building LoadTest client: %+v", err)
-	}
-	configureFunc(loadTestClient.Client)
-
 	loadTestsClient, err := loadtests.NewLoadTestsClientWithBaseURI(sdkApi)
 	if err != nil {
 		return nil, fmt.Errorf("building LoadTests client: %+v", err)
@@ -39,7 +31,6 @@ func NewClientWithBaseURI(sdkApi sdkEnv.Api, configureFunc func(c *resourcemanag
 	configureFunc(quotasClient.Client)
 
 	return &Client{
-		LoadTest:  loadTestClient,
 		LoadTests: loadTestsClient,
 		Quotas:    quotasClient,
 	}, nil
