--- resource-manager/loadtestservice/2022-12-01/loadtest/client.go
@@ -1,26 +1,0 @@
-package loadtest
-
-import (
-	"fmt"
-
-	"github.com/hashicorp/go-azure-sdk/sdk/client/resourcemanager"
-	sdkEnv "github.com/hashicorp/go-azure-sdk/sdk/environments"
-)
-
-// Copyright (c) Microsoft Corporation. All rights reserved.
-// Licensed under the MIT License. See NOTICE.txt in the project root for license information.
-
-type LoadTestClient struct {
-	Client *resourcemanager.Client
-}
-
-func NewLoadTestClientWithBaseURI(sdkApi sdkEnv.Api) (*LoadTestClient, error) {
-	client, err := resourcemanager.NewResourceManagerClient(sdkApi, "loadtest", defaultApiVersion)
-	if err != nil {
-		return nil, fmt.Errorf("instantiating LoadTestClient: %+v", err)
-	}
-
-	return &LoadTestClient{
-		Client: client,
-	}, nil
-}
