--- resource-manager/loadtestservice/2022-12-01/loadtest/id_loadtest_test.go
@@ -1,282 +1,0 @@
-package loadtest
-
-import (
-	"testing"
-
-	"github.com/hashicorp/go-azure-helpers/resourcemanager/resourceids"
-)
-
-// Copyright (c) Microsoft Corporation. All rights reserved.
-// Licensed under the MIT License. See NOTICE.txt in the project root for license information.
-
-var _ resourceids.ResourceId = &LoadTestId{}
-
-func TestNewLoadTestID(t *testing.T) {
-	id := NewLoadTestID("12345678-1234-9876-4563-123456789012", "example-resource-group", "loadTestValue")
-
-	if id.SubscriptionId != "12345678-1234-9876-4563-123456789012" {
-		t.Fatalf("Expected %q but got %q for Segment 'SubscriptionId'", id.SubscriptionId, "12345678-1234-9876-4563-123456789012")
-	}
-
-	if id.ResourceGroupName != "example-resource-group" {
-		t.Fatalf("Expected %q but got %q for Segment 'ResourceGroupName'", id.ResourceGroupName, "example-resource-group")
-	}
-
-	if id.LoadTestName != "loadTestValue" {
-		t.Fatalf("Expected %q but got %q for Segment 'LoadTestName'", id.LoadTestName, "loadTestValue")
-	}
-}
-
-func TestFormatLoadTestID(t *testing.T) {
-	actual := NewLoadTestID("12345678-1234-9876-4563-123456789012", "example-resource-group", "loadTestValue").ID()
-	expected := "/subscriptions/12345678-1234-9876-4563-123456789012/resourceGroups/example-resource-group/providers/Microsoft.LoadTestService/loadTests/loadTestValue"
-	if actual != expected {
-		t.Fatalf("Expected the Formatted ID to be %q but got %q", expected, actual)
-	}
-}
-
-func TestParseLoadTestID(t *testing.T) {
-	testData := []struct {
-		Input    string
-		Error    bool
-		Expected *LoadTestId
-	}{
-		{
-			// Incomplete URI
-			Input: "",
-			Error: true,
-		},
-		{
-			// Incomplete URI
-			Input: "/subscriptions",
-			Error: true,
-		},
-		{
-			// Incomplete URI
-			Input: "/subscriptions/12345678-1234-9876-4563-123456789012",
-			Error: true,
-		},
-		{
-			// Incomplete URI
-			Input: "/subscriptions/12345678-1234-9876-4563-123456789012/resourceGroups",
-			Error: true,
-		},
-		{
-			// Incomplete URI
-			Input: "/subscriptions/12345678-1234-9876-4563-123456789012/resourceGroups/example-resource-group",
-			Error: true,
-		},
-		{
-			// Incomplete URI
-			Input: "/subscriptions/12345678-1234-9876-4563-123456789012/resourceGroups/example-resource-group/providers",
-			Error: true,
-		},
-		{
-			// Incomplete URI
-			Input: "/subscriptions/12345678-1234-9876-4563-123456789012/resourceGroups/example-resource-group/providers/Microsoft.LoadTestService",
-			Error: true,
-		},
-		{
-			// Incomplete URI
-			Input: "/subscriptions/12345678-1234-9876-4563-123456789012/resourceGroups/example-resource-group/providers/Microsoft.LoadTestService/loadTests",
-			Error: true,
-		},
-		{
-			// Valid URI
-			Input: "/subscriptions/12345678-1234-9876-4563-123456789012/resourceGroups/example-resource-group/providers/Microsoft.LoadTestService/loadTests/loadTestValue",
-			Expected: &LoadTestId{
-				SubscriptionId:    "12345678-1234-9876-4563-123456789012",
-				ResourceGroupName: "example-resource-group",
-				LoadTestName:      "loadTestValue",
-			},
-		},
-		{
-			// Invalid (Valid Uri with Extra segment)
-			Input: "/subscriptions/12345678-1234-9876-4563-123456789012/resourceGroups/example-resource-group/providers/Microsoft.LoadTestService/loadTests/loadTestValue/extra",
-			Error: true,
-		},
-	}
-	for _, v := range testData {
-		t.Logf("[DEBUG] Testing %q", v.Input)
-
-		actual, err := ParseLoadTestID(v.Input)
-		if err != nil {
-			if v.Error {
-				continue
-			}
-
-			t.Fatalf("Expect a value but got an error: %+v", err)
-		}
-		if v.Error {
-			t.Fatal("Expect an error but didn't get one")
-		}
-
-		if actual.SubscriptionId != v.Expected.SubscriptionId {
-			t.Fatalf("Expected %q but got %q for SubscriptionId", v.Expected.SubscriptionId, actual.SubscriptionId)
-		}
-
-		if actual.ResourceGroupName != v.Expected.ResourceGroupName {
-			t.Fatalf("Expected %q but got %q for ResourceGroupName", v.Expected.ResourceGroupName, actual.ResourceGroupName)
-		}
-
-		if actual.LoadTestName != v.Expected.LoadTestName {
-			t.Fatalf("Expected %q but got %q for LoadTestName", v.Expected.LoadTestName, actual.LoadTestName)
-		}
-
-	}
-}
-
-func TestParseLoadTestIDInsensitively(t *testing.T) {
-	testData := []struct {
-		Input    string
-		Error    bool
-		Expected *LoadTestId
-	}{
-		{
-			// Incomplete URI
-			Input: "",
-			Error: true,
-		},
-		{
-			// Incomplete URI
-			Input: "/subscriptions",
-			Error: true,
-		},
-		{
-			// Incomplete URI (mIxEd CaSe since this is insensitive)
-			Input: "/sUbScRiPtIoNs",
-			Error: true,
-		},
-		{
-			// Incomplete URI
-			Input: "/subscriptions/12345678-1234-9876-4563-123456789012",
-			Error: true,
-		},
-		{
-			// Incomplete URI (mIxEd CaSe since this is insensitive)
-			Input: "/sUbScRiPtIoNs/12345678-1234-9876-4563-123456789012",
-			Error: true,
-		},
-		{
-			// Incomplete URI
-			Input: "/subscriptions/12345678-1234-9876-4563-123456789012/resourceGroups",
-			Error: true,
-		},
-		{
-			// Incomplete URI (mIxEd CaSe since this is insensitive)
-			Input: "/sUbScRiPtIoNs/12345678-1234-9876-4563-123456789012/rEsOuRcEgRoUpS",
-			Error: true,
-		},
-		{
-			// Incomplete URI
-			Input: "/subscriptions/12345678-1234-9876-4563-123456789012/resourceGroups/example-resource-group",
-			Error: true,
-		},
-		{
-			// Incomplete URI (mIxEd CaSe since this is insensitive)
-			Input: "/sUbScRiPtIoNs/12345678-1234-9876-4563-123456789012/rEsOuRcEgRoUpS/eXaMpLe-rEsOuRcE-GrOuP",
-			Error: true,
-		},
-		{
-			// Incomplete URI
-			Input: "/subscriptions/12345678-1234-9876-4563-123456789012/resourceGroups/example-resource-group/providers",
-			Error: true,
-		},
-		{
-			// Incomplete URI (mIxEd CaSe since this is insensitive)
-			Input: "/sUbScRiPtIoNs/12345678-1234-9876-4563-123456789012/rEsOuRcEgRoUpS/eXaMpLe-rEsOuRcE-GrOuP/pRoViDeRs",
-			Error: true,
-		},
-		{
-			// Incomplete URI
-			Input: "/subscriptions/12345678-1234-9876-4563-123456789012/resourceGroups/example-resource-group/providers/Microsoft.LoadTestService",
-			Error: true,
-		},
-		{
-			// Incomplete URI (mIxEd CaSe since this is insensitive)
-			Input: "/sUbScRiPtIoNs/12345678-1234-9876-4563-123456789012/rEsOuRcEgRoUpS/eXaMpLe-rEsOuRcE-GrOuP/pRoViDeRs/mIcRoSoFt.lOaDtEsTsErViCe",
-			Error: true,
-		},
-		{
-			// Incomplete URI
-			Input: "/subscriptions/12345678-1234-9876-4563-123456789012/resourceGroups/example-resource-group/providers/Microsoft.LoadTestService/loadTests",
-			Error: true,
-		},
-		{
-			// Incomplete URI (mIxEd CaSe since this is insensitive)
-			Input: "/sUbScRiPtIoNs/12345678-1234-9876-4563-123456789012/rEsOuRcEgRoUpS/eXaMpLe-rEsOuRcE-GrOuP/pRoViDeRs/mIcRoSoFt.lOaDtEsTsErViCe/lOaDtEsTs",
-			Error: true,
-		},
-		{
-			// Valid URI
-			Input: "/subscriptions/12345678-1234-9876-4563-123456789012/resourceGroups/example-resource-group/providers/Microsoft.LoadTestService/loadTests/loadTestValue",
-			Expected: &LoadTestId{
-				SubscriptionId:    "12345678-1234-9876-4563-123456789012",
-				ResourceGroupName: "example-resource-group",
-				LoadTestName:      "loadTestValue",
-			},
-		},
-		{
-			// Invalid (Valid Uri with Extra segment)
-			Input: "/subscriptions/12345678-1234-9876-4563-123456789012/resourceGroups/example-resource-group/providers/Microsoft.LoadTestService/loadTests/loadTestValue/extra",
-			Error: true,
-		},
-		{
-			// Valid URI (mIxEd CaSe since this is insensitive)
-			Input: "/sUbScRiPtIoNs/12345678-1234-9876-4563-123456789012/rEsOuRcEgRoUpS/eXaMpLe-rEsOuRcE-GrOuP/pRoViDeRs/mIcRoSoFt.lOaDtEsTsErViCe/lOaDtEsTs/lOaDtEsTvAlUe",
-			Expected: &LoadTestId{
-				SubscriptionId:    "12345678-1234-9876-4563-123456789012",
-				ResourceGroupName: "eXaMpLe-rEsOuRcE-GrOuP",
-				LoadTestName:      "lOaDtEsTvAlUe",
-			},
-		},
-		{
-			// Invalid (Valid Uri with Extra segment - mIxEd CaSe since this is insensitive)
-			Input: "/sUbScRiPtIoNs/12345678-1234-9876-4563-123456789012/rEsOuRcEgRoUpS/eXaMpLe-rEsOuRcE-GrOuP/pRoViDeRs/mIcRoSoFt.lOaDtEsTsErViCe/lOaDtEsTs/lOaDtEsTvAlUe/extra",
-			Error: true,
-		},
-	}
-	for _, v := range testData {
-		t.Logf("[DEBUG] Testing %q", v.Input)
-
-		actual, err := ParseLoadTestIDInsensitively(v.Input)
-		if err != nil {
-			if v.Error {
-				continue
-			}
-
-			t.Fatalf("Expect a value but got an error: %+v", err)
-		}
-		if v.Error {
-			t.Fatal("Expect an error but didn't get one")
-		}
-
-		if actual.SubscriptionId != v.Expected.SubscriptionId {
-			t.Fatalf("Expected %q but got %q for SubscriptionId", v.Expected.SubscriptionId, actual.SubscriptionId)
-		}
-
-		if actual.ResourceGroupName != v.Expected.ResourceGroupName {
-			t.Fatalf("Expected %q but got %q for ResourceGroupName", v.Expected.ResourceGroupName, actual.ResourceGroupName)
-		}
-
-		if actual.LoadTestName != v.Expected.LoadTestName {
-			t.Fatalf("Expected %q but got %q for LoadTestName", v.Expected.LoadTestName, actual.LoadTestName)
-		}
-
-	}
-}
-
-func TestSegmentsForLoadTestId(t *testing.T) {
-	segments := LoadTestId{}.Segments()
-	if len(segments) == 0 {
-		t.Fatalf("LoadTestId has no segments")
-	}
-
-	uniqueNames := make(map[string]struct{}, 0)
-	for _, segment := range segments {
-		uniqueNames[segment.Name] = struct{}{}
-	}
-	if len(uniqueNames) != len(segments) {
-		t.Fatalf("Expected the Segments to be unique but got %q unique segments and %d total segments", len(uniqueNames), len(segments))
-	}
-}
