--- resource-manager/loadtestservice/2022-12-01/loadtest/predicates.go
@@ -1,17 +1,0 @@
-package loadtest
-
-// Copyright (c) Microsoft Corporation. All rights reserved.
-// Licensed under the MIT License. See NOTICE.txt in the project root for license information.
-
-type OutboundEnvironmentEndpointOperationPredicate struct {
-	Category *string
-}
-
-func (p OutboundEnvironmentEndpointOperationPredicate) Matches(input OutboundEnvironmentEndpoint) bool {
-
-	if p.Category != nil && (input.Category == nil || *p.Category != *input.Category) {
-		return false
-	}
-
-	return true
-}
