+++ resource-manager/loadtestservice/2022-12-01/loadtests/README.md
@@ -100,13 +100,30 @@ for _, item := range items {
 ```
 
 
+### Example Usage: `LoadTestsClient.ListOutboundNetworkDependenciesEndpoints`
+
+```go
+ctx := context.TODO()
+id := loadtests.NewLoadTestID("12345678-1234-9876-4563-123456789012", "example-resource-group", "loadTestValue")
+
+// alternatively `client.ListOutboundNetworkDependenciesEndpoints(ctx, id)` can be used to do batched pagination
+items, err := client.ListOutboundNetworkDependenciesEndpointsComplete(ctx, id)
+if err != nil {
+	// handle the error
+}
+for _, item := range items {
+	// do something
+}
+```
+
+
 ### Example Usage: `LoadTestsClient.Update`
 
 ```go
 ctx := context.TODO()
 id := loadtests.NewLoadTestID("12345678-1234-9876-4563-123456789012", "example-resource-group", "loadTestValue")
 
-payload := loadtests.LoadTestResourcePatchRequestBody{
+payload := loadtests.LoadTestResourceUpdate{
 	// ...
 }
