+++ resource-manager/loadtestservice/2022-12-01/loadtests/method_delete.go
@@ -27,7 +27,6 @@ func (c LoadTestsClient) Delete(ctx context.Context, id LoadTestId) (result Dele
 		ExpectedStatusCodes: []int{
 			http.StatusAccepted,
 			http.StatusNoContent,
-			http.StatusOK,
 		},
 		HttpMethod: http.MethodDelete,
 		Path:       id.ID(),
