+++ resource-manager/loadtestservice/2022-12-01/loadtests/model_endpointdependency.go
@@ -1,10 +1,10 @@
-package loadtest
+package loadtests
 
 // Copyright (c) Microsoft Corporation. All rights reserved.
 // Licensed under the MIT License. See NOTICE.txt in the project root for license information.
 
 type EndpointDependency struct {
 	Description     *string           `json:"description,omitempty"`
-	DomainName      *string           `json:"domainName,omitempty"`
+	DomainName      string            `json:"domainName"`
 	EndpointDetails *[]EndpointDetail `json:"endpointDetails,omitempty"`
 }
