--- resource-manager/loadtestservice/2022-12-01/loadtests/model_loadtestresourcepatchrequestbody.go
@@ -1,14 +1,0 @@
-package loadtests
-
-import (
-	"github.com/hashicorp/go-azure-helpers/resourcemanager/identity"
-)
-
-// Copyright (c) Microsoft Corporation. All rights reserved.
-// Licensed under the MIT License. See NOTICE.txt in the project root for license information.
-
-type LoadTestResourcePatchRequestBody struct {
-	Identity   *identity.LegacySystemAndUserAssignedMap    `json:"identity,omitempty"`
-	Properties *LoadTestResourcePatchRequestBodyProperties `json:"properties,omitempty"`
-	Tags       *map[string]string                          `json:"tags,omitempty"`
-}
