+++ resource-manager/loadtestservice/2022-12-01/loadtests/model_loadtestresourceupdate.go
@@ -1,0 +1,14 @@
+package loadtests
+
+import (
+	"github.com/hashicorp/go-azure-helpers/resourcemanager/identity"
+)
+
+// Copyright (c) Microsoft Corporation. All rights reserved.
+// Licensed under the MIT License. See NOTICE.txt in the project root for license information.
+
+type LoadTestResourceUpdate struct {
+	Identity   *identity.LegacySystemAndUserAssignedMap `json:"identity,omitempty"`
+	Properties *LoadTestResourceUpdateProperties        `json:"properties,omitempty"`
+	Tags       *map[string]string                       `json:"tags,omitempty"`
+}
