+++ resource-manager/loadtestservice/2022-12-01/loadtests/model_loadtestresourceupdateproperties.go
@@ -3,7 +3,7 @@ package loadtests
 // Copyright (c) Microsoft Corporation. All rights reserved.
 // Licensed under the MIT License. See NOTICE.txt in the project root for license information.
 
-type LoadTestResourcePatchRequestBodyProperties struct {
+type LoadTestResourceUpdateProperties struct {
 	Description *string               `json:"description,omitempty"`
 	Encryption  *EncryptionProperties `json:"encryption,omitempty"`
 }
