+++ resource-manager/loadtestservice/2022-12-01/loadtests/model_outboundenvironmentendpoint.go
@@ -1,4 +1,4 @@
-package loadtest
+package loadtests
 
 // Copyright (c) Microsoft Corporation. All rights reserved.
 // Licensed under the MIT License. See NOTICE.txt in the project root for license information.
