+++ resource-manager/loadtestservice/2022-12-01/loadtests/predicates.go
@@ -30,3 +30,16 @@ func (p LoadTestResourceOperationPredicate) Matches(input LoadTestResource) bool
 
 	return true
 }
+
+type OutboundEnvironmentEndpointOperationPredicate struct {
+	Category *string
+}
+
+func (p OutboundEnvironmentEndpointOperationPredicate) Matches(input OutboundEnvironmentEndpoint) bool {
+
+	if p.Category != nil && (input.Category == nil || *p.Category != *input.Category) {
+		return false
+	}
+
+	return true
+}
