+++ resource-manager/loadtestservice/2022-12-01/quotas/model_checkquotaavailabilityresponse.go
@@ -8,9 +8,9 @@ import (
 // Licensed under the MIT License. See NOTICE.txt in the project root for license information.
 
 type CheckQuotaAvailabilityResponse struct {
-	Id         *string                                   `json:"id,omitempty"`
+	Id         string                                    `json:"id"`
 	Name       *string                                   `json:"name,omitempty"`
 	Properties *CheckQuotaAvailabilityResponseProperties `json:"properties,omitempty"`
 	SystemData *systemdata.SystemData                    `json:"systemData,omitempty"`
-	Type       *string                                   `json:"type,omitempty"`
+	Type       string                                    `json:"type"`
 }
