+++ resource-manager/loadtestservice/2022-12-01/quotas/model_quotabucketrequest.go
@@ -1,16 +1,8 @@
 package quotas
 
-import (
-	"github.com/hashicorp/go-azure-helpers/resourcemanager/systemdata"
-)
-
 // Copyright (c) Microsoft Corporation. All rights reserved.
 // Licensed under the MIT License. See NOTICE.txt in the project root for license information.
 
 type QuotaBucketRequest struct {
-	Id         *string                       `json:"id,omitempty"`
-	Name       *string                       `json:"name,omitempty"`
 	Properties *QuotaBucketRequestProperties `json:"properties,omitempty"`
-	SystemData *systemdata.SystemData        `json:"systemData,omitempty"`
-	Type       *string                       `json:"type,omitempty"`
 }
